+++ src/Timestepping/adaptivity.jl
@@ -37,35 +37,30 @@ struct NoAdaptivity <: AdaptivityCriterion end
 
 estimate_timestep(::NoAdaptivity, iter::AbstractSolver) = get_dt(iter)  # don't change current dt
 
-"""
+@doc raw"""
     AdaptBasedOnSegmentLength <: AdaptivityCriterion
     AdaptBasedOnSegmentLength(γ::Float64)
 
-Adapt timestep ``Δt`` based on the minimum distance ``ℓ_{\\min}`` between two filament nodes.
+Adapt timestep ``Δt`` based on the minimum distance ``δ`` between two filament nodes.
 
-More precisely, the timestep is set to ``Δt = γ ℓ_{\\min}² / β(ℓ_{\\min})``, where ``γ`` is a
-dimensionless factor to be chosen. Here ``β(ℓ) = Γ / 4π \\left[ \\ln(2ℓ_{\\min} / a) - Δ
-\\right]`` (see [`ParamsBiotSavart`](@ref) for the definitions of ``Γ``, ``a``
-and ``Δ``).
+More precisely, the timestep is set to ``Δt = γ T_{\text{KW}}(δ)``, where ``γ`` is a
+dimensionless factor to be chosen, and:
 
-Note that this criterion is based on simple dimensional analysis, assuming that
-``Δt`` must only depend on the minimum segment length ``ℓ_{\\min}`` and on the vortex
-circulation ``Γ`` (or rather on ``β``, which is proportional to ``Γ`` and also
-includes the effect of the vortex core size and profile).
+```math
+T_{\text{KW}}(λ) = \frac{2 λ^2}{Γ} {\left[
+    \ln\left( \frac{λ}{πa} \right) + \frac{1}{2} - (Δ + γ)
+\right]}^{-1}
+```
+
+is the period of a Kelvin wave of wavelength ``λ``.
+See [`ParamsBiotSavart`](@ref) for the definitions of the vortex parameters ``Γ``, ``a`` and
+``Δ``.
 
 This criterion is somewhat analogous to the CFL condition in grid-based
 computations, and ``γ`` is the analogous of the maximum CFL number to be allowed.
 As such, the right value of ``γ`` will depend on the chosen temporal scheme.
 
-Empirically, the following values of ``γ`` seem to provide stability and good precision:
-
-- ``γ = 1.8`` for [`RK4`](@ref),
-- ``γ = 1.2`` for [`SSPRK33`](@ref),
-- ``γ = 0.12`` for [`Euler`](@ref).
-
-This basically shows that among these methods, `RK4` should be preferred as it
-allows for a larger timestep for a given number of function evaluations. On the
-other hand, `Euler` should be always avoided!
+Empirically, the [`RK4`](@ref) scheme seems to require ``γ ≈ 1`` to remain stable.
 """
 struct AdaptBasedOnSegmentLength <: AdaptivityCriterion
     γ :: Float64
@@ -76,24 +71,21 @@ function estimate_timestep(crit::AdaptBasedOnSegmentLength, iter::AbstractSolver
     (; prob, fs,)  = iter
     p = prob.p :: ParamsBiotSavart
 
-    # 1. Determine minimum distance ℓ_min between filament nodes.
-    # TODO can we avoid iterating multiple times over the filaments?
-    T = eltype(p)
-    @assert T <: AbstractFloat
-    ℓ_min = convert(T, Inf)
-    @inbounds for f ∈ fs
-        ts = knots(f)
-        for i ∈ eachindex(segments(f))
-            ℓ::T = ts[i + 1] - ts[i]  # assume that knots roughly correspond to segment lengths
-            # @assert ℓ > 0
-            ℓ_min = min(ℓ_min, ℓ)
-        end
+    # 1. Determine minimum distance.
+    # If we're using the default filament parametrisation, then the knot increments are
+    # equal to the node distances (assuming knots were recently recomputed).
+    parametrisation = Filaments.parametrisation(fs)
+    δ = if parametrisation === Filaments.parametrise_by_node_distance
+        Filaments.minimum_knot_increment(fs)  # faster to evaluate
+    else
+        Filaments.minimum_node_distance(fs)
     end
 
     # 2. Determine timestep.
     (; Γ, a, Δ,) = p.common
-    β = (Γ / 4π) * (log(2 * ℓ_min / a) - Δ)
-    dt = γ * ℓ_min^2 / β
+    dt = 2 * γ * δ^2 / Γ / (
+        log(δ / (π * a)) + 1/2 - (Δ + MathConstants.γ)
+    )
 
     dt
 end